refactor: inject env data provider
This refactor take cares of the dataprovider package.
The data provider interface was change so `FetchData` will get the
resource type and the resource id as well and the return type is now
concrete type instead of an interface, and the `EnrichEvent` is now part
of the data provider interface.

The interesting part is in
(flavors/posture.go)[https://github.com/elastic/cloudbeat/pull/792/files#diff-fc3709e0663635e41188687fee0307f0194885d7c4af2d1b7cd3ff79833bf6f2R224]
where the data provider is been initiated. The refactor moved all the
client creation and all the required metadata to the root so that the
data provider so not require any non-static dependency