Fetch AWSSupportAccess policy only once (#900)
`GetPolicies()` conflicts with `GetSupportPolicy()` because they both
create a `AWSSupportAccess` resource with subtype `aws-policy`. The
`GetSupportPolicy` one has the roles array and the one from
`GetPolicies` doesn't (set to `null`). Thus, the first one passes 1.17
and the second one doesn't.

Here is a fix which produces only one of these resources, we might want
to eventually revisit this approach but with this change at least fixes
the problematic behavior without affecting non-functional behavior.