Azure ARM: Use temporary public key and disable SSH (#1363)
This uses the script to kill SSH as an additional precaution.

Public IP to be removed in future change.

- Generate public ssh key:
This removes the need for the seed parameter by using a
`Microsoft.Resources/deploymentScripts` resource to generate a temporary
public key. The private key is thrown out and ssh is still disabled by
the deployment script.

Downside of this approach is that the deployment script needs some time
to complete.