Add Regal Rego linting to project (#1513)
Hello elastic friends! 👋😃 Originally submitted to the CSP repo, but
I was instructed to resubmit the PR here.

[Regal](https://github.com/StyraInc/regal)
is a linter for Rego, and it's reaching a level of maturity where
I feel confident putting it to use in some serious policy libraries.

Besides linting, the CI integration helps annotate PRs with violations
at the location they happened, helping developers quickly figure out what
they need to change.

As part of integrating the linter in this repo, I went ahead and fixed a
number of issues reported. Naturally, I've made sure that all the Rego
tests pass, but if there's any change you feel was too invasive, or perhaps
changed code that some other project depend on, do let me know and I'll be
happy to revert that.

Issues fixed for the following rules:

* [not-equals-in-loop](https://docs.styra.com/regal/rules/bugs/not-equals-in-loop)
* [unconditional-assignment](https://docs.styra.com/regal/rules/style/unconditional-assignment)
* [use-in-operator](https://docs.styra.com/regal/rules/idiomatic/use-in-operator)
* [redundant-alias](https://docs.styra.com/regal/rules/imports/redundant-alias)
* [no-whitespace-comment](https://docs.styra.com/regal/rules/style/no-whitespace-comment)
* [constant-condition](https://docs.styra.com/regal/rules/bugs/constant-condition)
* [non-raw-regex-pattern](https://docs.styra.com/regal/rules/idiomatic/non-raw-regex-pattern)
* [redundant-alias](https://docs.styra.com/regal/rules/imports/redundant-alias)

Many issues, primarily related to style, have been disabled in the Regal config
for now, and I'm sure you'll want to work on having some of them enabled later.

Signed-off-by: Anders Eknert <anders@styra.com>