update-hermit: Consider major versions as well
`hermit upgrade` will only upgrade minor versions for packages that use
semantic versioning. Not all of our hermit packages use semantic
versioning and even for those that do, we want to be at least aware of
new major versions and update to those if they don't cause too much of a
breakage.
Uninstalling and re-installing will always install to the latest major
version.

This fixes an issue with `pre-commit autoupdate` not being able to
update our `black` hook because it's too old to understand newer syntax.

Additionally, skip the poetry-based checks in pre-commit as that is not
installed.