Azure ARM: Use resource group deployment schema (#1624)
This uses the "default" deployment template which includes a better UI for
selecting resource group to deploy to. This also removes one sub-deployment for
the resource group as the user can select/create one from the UI.

Previously, it was thought that we need the subscription deployment template
schema to deploy role assignments to the subscription level.

Removes the location parameter and use the native Azure selector. Additionally,
fixes an issue with the deployment not working across different regions
(#1619). Fix is just to include the location in the name for the role
assignment deployment. This one is the only sub-deployment that needs to be
fixed because the rest deploy at the resource group level. The subscription
level sub-deployment requires the `location` property, otherwise you get an
error:
```
The 'location' property must be specified for
'role-assignment-deployment-eastus'. Please see
https://aka.ms/arm-deployment-subscription for usage details. (Code:
InvalidDeployment)
```

Finally, I've fixed ci commands that were using the old filename for the
template.

(cherry picked from commit 4180191e84f706f9695519f00e341887f360bad8)