Azure: Fix wrong usage of subscription ID
There are two forms of subscription IDs:
- Short: e.g. 8d65815f-a5b6-402f-9298-045155da7d74
- Fully Qualified: e.g. /subscriptions/8d65815f-a5b6-402f-9298-045155da7d74

The batch fetcher was grouping assets using the first one and then was
using the second one to access the keys from the group map.

This commit fixes that and also re-names the fields to avoid future
confusion. A `ShortId` field has also been added to be used for
accessing the group maps and formatting batch asset names.