revive: Enable unexported-naming and unexported-return
These are also flagged by Goland so it's nice to have

unexported-naming

Description: this rule warns on wrongly named un-exported symbols, i.e. un-exported symbols whose name start with a capital letter.

Configuration: N/A

unexported-return

Description: This rule warns when an exported function or method returns a value of an un-exported type.

Configuration: N/A