revive: Enable cognitive-complexity and cyclomatic (#1720)
cognitive-complexity: Start with a value of 30 instead of the default 7 which I find too restrictive.
cyclomatic: Start with 15 instead of default 10.

I've excluded all existing functions that fail the checks since refactoring them is a lot of work. The benefit of enabling the rules is that it's going to fail for hard-to-read functions in future code reviews.

Description: Cognitive complexity is a measure of how hard code is to understand. While cyclomatic complexity is good to measure "testability" of the code, cognitive complexity aims to provide a more precise measure of the difficulty of understanding the code. Enforcing a maximum complexity per function helps to keep code readable and maintainable.

Configuration: (int) the maximum function complexity

https://www.sonarsource.com/docs/CognitiveComplexity.pdf

cyclomatic

Description: Cyclomatic complexity is a measure of code complexity. Enforcing a maximum complexity per function helps to keep code readable and maintainable.

Configuration: (int) the maximum function complexity