Azure: Deploy using dev script (#1689)
Add script to deploy azure ARM template with ssh enabled for the VMs.

The generated templates have been added to the repo since they are not affected by deployment parameters. Running the script is not necessary to deploy the the templates, it's also possible to do that manually from the azure cli or the UI.

A pre-commit action has been added to make sure the scripts work after changes to the ARM templates.

The script works for both the organization and single-account templates.