Temporarily disable long running tests
Before this change all the PRs were executing these long running jobs, on avg ~15min each run. But the branch with the changes was not the one being tested, rather just main, therefore wasting time and resource. With the ticket https://github.com/elastic/security-team/issues/9129 we will solve this issue. But in the meanwhile, we are disabling these jobs.